Translate control deck inputs to Xbox/DS4
Works with no apparent latency and negligible CPU usage.

Only works with one of the control deck controllers at a time.
Using two results in the virtual controller inputs getting
stuck and the devices as a whole becoming unresponsive.

There is a similar effect when enabling the DS4 PlayStation
button. It's therefore commented out in this commit.

The code is intentionally verbose and avoids any simplification
possible with Python in case performance turns out to be
inadequate and it has to be ported to C++.